+++ src/mixnorm.c
@@ -15,7 +15,7 @@ mcmclib_mixnorm_lpdf* mcmclib_mixnorm_lpdf_alloc(gsl_vector* w, mcmclib_mvnorm_l
   mcmclib_mixnorm_lpdf* a = (mcmclib_mixnorm_lpdf*) malloc(sizeof(mcmclib_mixnorm_lpdf));
   a->pis = (mcmclib_mvnorm_lpdf**) malloc(w->size * sizeof(mcmclib_mvnorm_lpdf*));
   memcpy(a->pis, pis, w->size * sizeof(mcmclib_mvnorm_lpdf*));
-  for(int i=0; i<w->size; i++)
+  for(size_t i=0; i<w->size; i++)
     a->pis[i] = pis[i];
   a->w = gsl_vector_alloc(w->size);
   gsl_vector_memcpy(a->w, w);
@@ -31,7 +31,7 @@ void mcmclib_mixnorm_lpdf_free(mcmclib_mixnorm_lpdf* p) {
 double mcmclib_mixnorm_lpdf_compute(void* in_p, gsl_vector* x) {
   mcmclib_mixnorm_lpdf* p = (mcmclib_mixnorm_lpdf*) in_p;
   double ans = 0.0;
-  for(int i = 0; i < p->w->size; i++)
+  for(size_t i = 0; i < p->w->size; i++)
     ans += exp(mcmclib_mvnorm_lpdf_compute(p->pis[i], x)) * gsl_vector_get(p->w, i);
   return log(ans);
 }
